pbdSLAP-package {pbdSLAP}	R Documentation

Programming with Big Data – Scalable Linear Algebra Packages

Description

pbdSLAP utilizes scalable linear algebra packages mainly including BLACS, PBLAS, and ScaLAPACK in double precision via pbdMPI based on ScaLAPACK version 2.0.2.

Details

Package: 	 pbdSLAP

Type: 	 Package

License: 	 MPL2

LazyLoad: 	 yes

This package requires pbdMPI and MPI system. The main purpose of pbdSLAP is to provide several scalable linear algebra packages containing double precision libraries for pbdDMAC or other useful packages.

Author(s)

Wei-Chen Chen wccsnow@gmail.com, Drew Schmidt, George Ostrouchov, and Pragneshkumar Patel.

References

Programming with Big Data in R Website: http://r-pbd.org/

ScaLAPACK Website: http://www.netlib.org/scalapack/

ScaLAPACK Block Cyclic Data Distribution Website: http://acts.nersc.gov/scalapack/hands-on/datadist.html

Examples

## Not run: 
### Under command mode, run the demo with 2 processors by
### (Use Rscript.exe for windows system)

mpiexec -np 2 Rscript -e "demo(gridinfo,'pbdSLAP',ask=F,echo=F)"

## End(Not run)

[Package pbdSLAP version 0.2-1 Index]