SLAP Grid Information {pbdSLAP}	R Documentation

SLAP Grid

Description

These functions initializes a grid of pbdSLAP, assigns the information to a global object, and free the grid.

Usage

  slap.init.grid(nprow, npcol = 1, ictxt = 0)
  slap.exit.grid(ictxt)
  slap.finalize(quit.mpi = FALSE)

Arguments

nprow
number of row processors.

npcol
number of column processors.

ictxt
the grid id

quit.mpi
if finalize MPI.

Details

This function arranges all processors in a (nprow * npcol) grid and the grid will map the big data matrix.

Value

slap.init.grid assigns a global object .__grid_info_0 for ictxt = 0 containing the grid information. slap.exit.grid free the grid. slap.finalize free all memory.

Author(s)

Wei-Chen Chen wccsnow@gmail.com, Drew Schmidt, George Ostrouchov, and Pragneshkumar Patel.

References

Programming with Big Data in R Website: http://r-pbd.org/

ScaLAPACK Website: http://www.netlib.org/scalapack/

ScaLAPACK Block Cyclic Data Distribution Website: http://acts.nersc.gov/scalapack/hands-on/datadist.html

Examples

## Not run: 
### Under command mode, run the demo with 2 processors by
### (Use Rscript.exe for windows system)

mpiexec -np 2 Rscript -e "demo(gridinfo,'pbdSLAP',ask=F,echo=F)"

## End(Not run)

[Package pbdSLAP version 0.2-1 Index]